Move mapper to AdyenCheckout
